requirements.txt
# ── Attention Monitor — Flask Server Dependencies ──
# Install with: pip install -r requirements.txt

# Web server
flask==2.3.3
flask-cors==4.0.0

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78   # required for cv2.face (LBF landmark model)
Pillow==10.1.0
numpy==1.24.4

# Machine Learning
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.0.3

# NLP / Transformers
torch==2.1.0
transformers==4.35.2
tokenizers==0.15.0

# Utilities
scipy==1.11.4
